# blog-server/docs/swagger.yaml
basePath: /
definitions:
  advert_api.AdvertRequest:
    properties:
      href:
        description: 跳转链接
        type: string
      images:
        description: 图片
        type: string
      is_show:
        description: 是否提示
        type: boolean
      title:
        description: 显示的标题
        type: string
    required:
    - href
    - images
    - title
    type: object
  ctype.ImageType:
    enum:
    - 1
    - 2
    type: integer
    x-enum-comments:
      Local: 本地
      QiNiu: 七牛云
    x-enum-varnames:
    - Local
    - QiNiu
  image_ser.FileUploadResponse:
    properties:
      file_name:
        type: string
      is_success:
        description: 图片是否上传成功
        type: boolean
      msg:
        type: string
    type: object
  images_api.ImageResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      path:
        type: string
    type: object
  images_api.ImageUpdateRequest:
    properties:
      id:
        type: integer
      name:
        type: string
    required:
    - id
    - name
    type: object
  menu_api.Banner:
    properties:
      id:
        type: integer
      path:
        type: string
    type: object
  menu_api.ImageSort:
    properties:
      image_id:
        type: integer
      sort:
        type: integer
    type: object
  menu_api.MenuNameResponse:
    properties:
      id:
        type: integer
      path:
        type: string
      title:
        type: string
    type: object
  menu_api.MenuResponse:
    properties:
      abstract:
        description: 简介
        items:
          type: string
        type: array
      abstract_time:
        description: 简介的切换时间
        type: integer
      banner_time:
        description: 菜单图片的切换时间，为0表示不切换
        type: integer
      banners:
        items:
          $ref: '#/definitions/menu_api.Banner'
        type: array
      created_at:
        description: 创建时间
        type: string
      id:
        description: 主键id
        type: integer
      path:
        type: string
      slogan:
        type: string
      sort:
        description: 菜单的顺序
        type: integer
      title:
        type: string
      update_at:
        description: 更新时间
        type: string
    type: object
  menu_api.menuRequest:
    properties:
      abstract:
        items:
          type: string
        type: array
      abstract_time:
        description: 切换的时间
        type: integer
      banner_time:
        description: 切换的时间
        type: integer
      image_sort_list:
        description: 具体图片的顺序
        items:
          $ref: '#/definitions/menu_api.ImageSort'
        type: array
      path:
        type: string
      slogan:
        type: string
      sort:
        description: 菜单的序号
        type: integer
      title:
        type: string
    required:
    - path
    - sort
    - title
    type: object
  models.AdvertModel:
    properties:
      created_at:
        description: 创建时间
        type: string
      href:
        description: 跳转链接
        type: string
      id:
        description: 主键id
        type: integer
      images:
        description: 图片
        type: string
      is_show:
        description: 是否提示
        type: boolean
      title:
        description: 显示的标题
        type: string
      update_at:
        description: 更新时间
        type: string
    type: object
  models.BannerModel:
    properties:
      created_at:
        description: 创建时间
        type: string
      hash:
        description: 图片的hash值，用于判断重复图片
        type: string
      id:
        description: 主键id
        type: integer
      image_type:
        allOf:
        - $ref: '#/definitions/ctype.ImageType'
        description: 图片类型，本地还是七牛
      name:
        description: 图片名称
        type: string
      path:
        description: 图片路径
        type: string
      update_at:
        description: 更新时间
        type: string
    type: object
  models.RemoveRequest:
    properties:
      id_list:
        items:
          type: integer
        type: array
    type: object
  res.ListResponse-models_AdvertModel:
    properties:
      count:
        type: integer
      list:
        $ref: '#/definitions/models.AdvertModel'
    type: object
  res.ListResponse-models_BannerModel:
    properties:
      count:
        type: integer
      list:
        $ref: '#/definitions/models.BannerModel'
    type: object
  res.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: 127.0.0.01:8080
info:
  contact: {}
  description: API文档
  title: gvb_server API文档
  version: "1.0"
paths:
  /api/adverts:
    delete:
      description: 批量删除广告
      parameters:
      - description: 广告id列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.RemoveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/res.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 批量删除广告
      tags:
      - 广告管理
    get:
      description: 广告列表
      parameters:
      - in: query
        name: key
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/res.Response'
            - properties:
                data:
                  $ref: '#/definitions/res.ListResponse-models_AdvertModel'
              type: object
      summary: 广告列表
      tags:
      - 广告管理
    post:
      description: 创建广告
      parameters:
      - description: 表示多个参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/advert_api.AdvertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
      summary: 创建广告
      tags:
      - 广告管理
  /api/adverts/:id:
    put:
      description: 更新广告
      parameters:
      - description: 广告的一些参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/advert_api.AdvertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/res.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 更新广告
      tags:
      - 广告管理
  /api/image_names:
    get:
      description: 图片名称列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/res.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/images_api.ImageResponse'
                  type: array
              type: object
      summary: 图片名称列表
      tags:
      - 图片管理
  /api/images:
    delete:
      description: 批量删除图片
      parameters:
      - description: 图片id列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.RemoveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/res.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 批量删除图片
      tags:
      - 图片管理
    get:
      description: 图片列表
      parameters:
      - in: query
        name: key
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/res.Response'
            - properties:
                data:
                  $ref: '#/definitions/res.ListResponse-models_BannerModel'
              type: object
      summary: 图片列表
      tags:
      - 图片管理
    post:
      description: 图片上传
      parameters:
      - description: 本应使用表单多图片文件参数
        in: body
        name: data
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/res.Response'
            - properties:
                data:
                  $ref: '#/definitions/image_ser.FileUploadResponse'
              type: object
      summary: 图片上传
      tags:
      - 图片管理
    put:
      description: 图片更新
      parameters:
      - description: 图片修改内容
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/images_api.ImageUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/res.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 图片更新
      tags:
      - 图片管理
  /api/menu_names:
    get:
      description: 菜单名称列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/res.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/menu_api.MenuNameResponse'
                  type: array
              type: object
      summary: 菜单名称列表
      tags:
      - 菜单管理
  /api/menus:
    delete:
      description: 菜单删除
      parameters:
      - description: 删除菜单
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.RemoveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/res.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 菜单删除
      tags:
      - 菜单管理
    get:
      description: 菜单列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/res.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/menu_api.MenuResponse'
                  type: array
              type: object
      summary: 菜单列表
      tags:
      - 菜单管理
    post:
      description: 添加菜单
      parameters:
      - description: 添加菜单
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/menu_api.menuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/res.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 添加菜单
      tags:
      - 菜单管理
  /api/menus/:id:
    get:
      description: 菜单详情
      parameters:
      - description: 菜单详情
        in: body
        name: data
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/res.Response'
            - properties:
                data:
                  $ref: '#/definitions/menu_api.MenuResponse'
              type: object
      summary: 菜单详情
      tags:
      - 菜单管理
    put:
      description: 菜单更新
      parameters:
      - description: 菜单更新
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/menu_api.menuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/res.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 菜单更新
      tags:
      - 菜单管理
  /api/settings/:name:
    get:
      description: 配置信息
      parameters:
      - in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 数据类型应为对应数据的结构体
          schema:
            allOf:
            - $ref: '#/definitions/res.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 配置信息
      tags:
      - 系统配置
    put:
      description: 修改配置信息
      parameters:
      - in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 数据类型应为对应数据的结构体
          schema:
            allOf:
            - $ref: '#/definitions/res.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 修改配置信息
      tags:
      - 系统配置
swagger: "2.0"